Fix go module name, refactor project structure accordingly
