+++ .gitignore
@@ -1,4 +1,4 @@
-scfw
+/bin/
 .oss-readiness
 dist/
