+++ scfw/internal/cli/root_test.go
