+++ scfw/internal/ddapi/auth.go
