+++ scfw/internal/ddapi/auth_test.go
