+++ scfw/internal/ecosystem/npm_test.go
