+++ scfw/internal/pm/npm/npm_temp_project.go
@@ -20,8 +20,8 @@ import (
 	"slices"
 	"strings"
 
-	"github.com/DataDog/scfw/scfw/internal/ecosystem"
-	"github.com/DataDog/scfw/scfw/internal/pm"
+	"github.com/DataDog/supply-chain-firewall/scfw/internal/ecosystem"
+	"github.com/DataDog/supply-chain-firewall/scfw/internal/pm"
 )
 
 // npmFileURIPrefix is the URI scheme npm uses for local package dependencies.
